Rename: 좀더 범용적인 exception naming
