+++ src/main/java/com/project/doubleshop/web/common/StatusRequest.java
@@ -1,4 +1,4 @@
-package com.project.doubleshop.web.item.dto;
+package com.project.doubleshop.web.common;
 
 import com.project.doubleshop.domain.common.Status;
 
@@ -10,11 +10,11 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @AllArgsConstructor
-public class ItemStatusRequest {
+public class StatusRequest {
 	Long id;
 	Status status;
 
-	public ItemStatusRequest(Status status) {
+	public StatusRequest(Status status) {
 		this.status = status;
 	}
 }
