+++ src/main/java/com/project/doubleshop/web/item/exception/DataNotFoundException.java
@@ -1,0 +1,24 @@
+package com.project.doubleshop.web.item.exception;
+
+public class DataNotFoundException extends RuntimeException {
+	public DataNotFoundException() {
+		super();
+	}
+
+	public DataNotFoundException(String message) {
+		super(message);
+	}
+
+	public DataNotFoundException(String message, Throwable cause) {
+		super(message, cause);
+	}
+
+	public DataNotFoundException(Throwable cause) {
+		super(cause);
+	}
+
+	protected DataNotFoundException(String message, Throwable cause, boolean enableSuppression,
+		boolean writableStackTrace) {
+		super(message, cause, enableSuppression, writableStackTrace);
+	}
+}
